+++ priv/repo/migrations/20240124054202_add_wiped_to_dataclips.exs
@@ -1,0 +1,13 @@
+defmodule Lightning.Repo.Migrations.AddWipedToDataclips do
+  use Ecto.Migration
+
+  def change do
+    alter table("dataclips") do
+      add :wiped_at, :utc_datetime
+    end
+
+    alter table("steps") do
+      modify :input_dataclip_id, :binary_id, null: true, from: {:binary_id, null: false}
+    end
+  end
+end