+++ test/lightning_web/channels/run_channel_test.exs
@@ -1,4 +1,5 @@
 defmodule LightningWeb.RunChannelTest do
+  alias Lightning.Invocation.Dataclip
   use LightningWeb.ChannelCase
 
   alias Lightning.Invocation.Step
@@ -179,7 +180,66 @@ defmodule LightningWeb.RunChannelTest do
                "jobs" => jobs,
                "edges" => edges,
                "starting_node_id" => run.starting_trigger_id,
-               "dataclip_id" => run.dataclip_id
+               "dataclip_id" => run.dataclip_id,
+               "options" => %{"output_dataclips" => true}
+             }
+    end
+
+    test "fetch:plan for project with erase_all retention setting", %{
+      credential: credential
+    } do
+      project = insert(:project, retention_policy: :erase_all)
+
+      %{run: run, workflow: workflow} =
+        create_run(%{project: project, credential: credential})
+
+      %{socket: socket} = create_socket(%{run: run})
+
+      id = run.id
+      ref = push(socket, "fetch:plan", %{})
+
+      assert_reply ref, :ok, payload
+
+      triggers =
+        workflow.triggers
+        |> Enum.map(&Map.take(&1, [:id]))
+        |> Enum.map(&stringify_keys/1)
+
+      [job] = workflow.jobs
+
+      jobs =
+        [
+          %{
+            "id" => job.id,
+            "name" => job.name,
+            "body" => job.body,
+            "credential_id" => credential.id,
+            "adaptor" => "@openfn/language-common@1.6.2"
+          }
+        ]
+
+      edges =
+        workflow.edges
+        |> Enum.map(
+          &(Map.take(&1, [
+              :id,
+              :source_trigger_id,
+              :source_job_id,
+              :enabled,
+              :target_job_id
+            ])
+            |> Map.put(:condition, "state.a == 33"))
+        )
+        |> Enum.map(&stringify_keys/1)
+
+      assert payload == %{
+               "id" => id,
+               "triggers" => triggers,
+               "jobs" => jobs,
+               "edges" => edges,
+               "starting_node_id" => run.starting_trigger_id,
+               "dataclip_id" => run.dataclip_id,
+               "options" => %{"output_dataclips" => false}
              }
     end
 
@@ -207,6 +267,45 @@ defmodule LightningWeb.RunChannelTest do
       ref = push(socket, "fetch:dataclip", %{})
 
       assert_reply ref, :ok, {:binary, ~s<{"foo": "bar"}>}
+    end
+
+    test "fetch:dataclip wipes dataclip body for projects with erase_all retention policy",
+         %{
+           credential: credential
+         } do
+      # erase_all
+      project = insert(:project, retention_policy: :erase_all)
+
+      %{run: run, dataclip: dataclip} =
+        create_run(%{project: project, credential: credential})
+
+      %{socket: socket} = create_socket(%{run: run})
+
+      ref = push(socket, "fetch:dataclip", %{})
+
+      assert_reply ref, :ok, {:binary, _payload}
+
+      # dataclip body is cleared
+      updated_dataclip = Lightning.Invocation.get_dataclip_details!(dataclip.id)
+      assert updated_dataclip.wiped_at
+      refute updated_dataclip.body
+
+      # retain_all
+      project = insert(:project, retention_policy: :retain_all)
+
+      %{run: run, dataclip: dataclip} =
+        create_run(%{project: project, credential: credential})
+
+      %{socket: socket} = create_socket(%{run: run})
+
+      ref = push(socket, "fetch:dataclip", %{})
+
+      assert_reply ref, :ok, {:binary, _payload}
+
+      # dataclip body is not cleared
+      updated_dataclip = Lightning.Invocation.get_dataclip_details!(dataclip.id)
+      refute updated_dataclip.wiped_at
+      assert updated_dataclip.body
     end
 
     test "fetch:credential", %{socket: socket, credential: credential} do
@@ -379,17 +478,49 @@ defmodule LightningWeb.RunChannelTest do
         run: run,
         user: user,
         workflow: workflow,
-        credential: credential
+        credential: credential,
+        project: project
       }
     end
 
     test "step:start", %{
       socket: socket,
-      run: run,
+      run: %{dataclip_id: dataclip_id},
       workflow: workflow,
-      credential: %{id: credential_id}
+      credential: %{id: credential_id} = credential,
+      project: project
     } do
       # { id, job_id, input_dataclip_id }
+      step_id = Ecto.UUID.generate()
+      [%{id: job_id}] = workflow.jobs
+
+      ref =
+        push(socket, "step:start", %{
+          "step_id" => step_id,
+          "credential_id" => credential_id,
+          "job_id" => job_id,
+          "input_dataclip_id" => dataclip_id
+        })
+
+      assert_reply ref, :ok, %{step_id: ^step_id}, 1_000
+
+      assert project.retention_policy == :retain_all
+
+      assert %{
+               credential_id: ^credential_id,
+               job_id: ^job_id,
+               input_dataclip_id: ^dataclip_id
+             } =
+               Repo.get!(Step, step_id)
+
+      # input dataclip is not saved for projects with erase_all
+      project = insert(:project, retention_policy: :erase_all)
+
+      %{run: run, workflow: workflow} =
+        create_run(%{project: project, credential: credential})
+
+      %{socket: socket} = create_socket(%{run: run})
+
       step_id = Ecto.UUID.generate()
       [%{id: job_id}] = workflow.jobs
 
@@ -401,9 +532,15 @@ defmodule LightningWeb.RunChannelTest do
           "input_dataclip_id" => run.dataclip_id
         })
 
-      assert_reply ref, :ok, %{step_id: ^step_id}, 1_000
+      assert_reply ref, :ok, %{step_id: ^step_id}
 
-      assert %{credential_id: ^credential_id, job_id: ^job_id} =
+      assert project.retention_policy == :erase_all
+
+      assert %{
+               credential_id: ^credential_id,
+               job_id: ^job_id,
+               input_dataclip_id: nil
+             } =
                Repo.get!(Step, step_id)
     end
 
@@ -445,6 +582,61 @@ defmodule LightningWeb.RunChannelTest do
 
       assert_reply ref, :ok, %{step_id: ^step_id}
       assert %{exit_reason: "fail"} = Repo.get(Step, step.id)
+    end
+
+    test "step:complete does not save the dataclip if project retention policy is set to erase_all",
+         %{
+           socket: socket,
+           run: run,
+           workflow: workflow,
+           project: project,
+           credential: credential
+         } do
+      [job] = workflow.jobs
+      %{id: step_id} = insert(:step, runs: [run], job: job)
+      dataclip_id = Ecto.UUID.generate()
+
+      ref =
+        push(socket, "step:complete", %{
+          "step_id" => step_id,
+          "output_dataclip_id" => dataclip_id,
+          "output_dataclip" => ~s({"foo": "bar"}),
+          "reason" => "normal"
+        })
+
+      assert_reply ref, :ok, %{step_id: ^step_id}
+
+      # dataclip is saved
+      assert project.retention_policy == :retain_all
+      assert %{output_dataclip_id: ^dataclip_id} = Repo.get(Step, step_id)
+      assert Repo.get(Dataclip, dataclip_id)
+
+      # project with erase_all
+      project = insert(:project, retention_policy: :erase_all)
+
+      %{run: run, workflow: workflow} =
+        create_run(%{project: project, credential: credential})
+
+      %{socket: socket} = create_socket(%{run: run})
+
+      [job] = workflow.jobs
+      %{id: step_id} = insert(:step, runs: [run], job: job)
+      dataclip_id = Ecto.UUID.generate()
+
+      ref =
+        push(socket, "step:complete", %{
+          "step_id" => step_id,
+          "output_dataclip_id" => dataclip_id,
+          "output_dataclip" => ~s({"foo": "bar"}),
+          "reason" => "normal"
+        })
+
+      assert_reply ref, :ok, %{step_id: ^step_id}
+
+      # dataclip is NOT saved
+      assert project.retention_policy == :erase_all
+      assert %{output_dataclip_id: nil} = Repo.get(Step, step_id)
+      refute Repo.get(Dataclip, dataclip_id)
     end
   end
 
@@ -800,6 +992,14 @@ defmodule LightningWeb.RunChannelTest do
   defp create_socket_and_run(%{credential: credential, user: user}) do
     project = insert(:project, project_users: [%{user: user}])
 
+    run_result = create_run(%{project: project, credential: credential})
+
+    socket_result = create_socket(run_result)
+
+    Map.merge(run_result, socket_result)
+  end
+
+  defp create_run(%{project: project, credential: credential}) do
     dataclip =
       insert(:http_request_dataclip, project: project)
 
@@ -813,7 +1013,7 @@ defmodule LightningWeb.RunChannelTest do
 
     workflow =
       %{triggers: [trigger]} =
-      build(:workflow)
+      build(:workflow, project: project)
       |> with_trigger(trigger)
       |> with_job(job)
       |> with_edge({trigger, job}, %{
@@ -836,6 +1036,15 @@ defmodule LightningWeb.RunChannelTest do
         dataclip: dataclip
       )
 
+    %{
+      run: run,
+      workflow: workflow,
+      credential: credential,
+      dataclip: dataclip
+    }
+  end
+
+  defp create_socket(%{run: run}) do
     Lightning.Stub.reset_time()
 
     {:ok, bearer, _} =
@@ -853,13 +1062,7 @@ defmodule LightningWeb.RunChannelTest do
         %{"token" => Workers.generate_run_token(run)}
       )
 
-    %{
-      socket: socket,
-      run: run,
-      workflow: workflow,
-      credential: credential,
-      dataclip: dataclip
-    }
+    %{socket: socket}
   end
 
   defp stringify_keys(map) do