+++ test/lightning_web/live/project_live_test.exs
@@ -1754,7 +1754,7 @@ defmodule LightningWeb.ProjectLiveTest do
     end
   end
 
-  describe "data-retention" do
+  describe "data-storage" do
     setup :register_and_log_in_user
     setup :create_project_for_current_user
 
@@ -1763,10 +1763,10 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, _view, html} =
         live(
           conn,
-          ~p"/projects/#{project.id}/settings#data-retention"
+          ~p"/projects/#{project.id}/settings#data-storage"
         )
 
-      assert html =~ "Workflow Input &amp; Output Data Retention"
+      assert html =~ "Input/Output Data Storage Policy"
     end
 
     @tag role: :admin
@@ -1774,13 +1774,11 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, html} =
         live(
           conn,
-          ~p"/projects/#{project.id}/settings#data-retention"
+          ~p"/projects/#{project.id}/settings#data-storage"
         )
 
-      assert html =~ "Workflow Input &amp; Output Data Retention"
-
-      assert html =~
-               "Determine whether Input and Output data for workflow should be retained or erased."
+      assert html =~ "Input/Output Data Storage Policy"
+      assert html =~ "Should OpenFn store input/output data for workflow runs?"
 
       # retain_all is the default
       assert ["checked"] ==
@@ -1808,7 +1806,7 @@ defmodule LightningWeb.ProjectLiveTest do
       refute html =~ "heads-up-description"
 
       # 3 radio buttons descriptions
-      assert "Retain all input & output data on workflow completion" =
+      assert "Retain input/output data for all workflow runs" =
                view
                |> element(~s{label#[for="retain_all"]})
                |> render()
@@ -1816,7 +1814,7 @@ defmodule LightningWeb.ProjectLiveTest do
                |> Floki.text()
                |> String.trim()
 
-      assert "Retain input & output on workflow run attempts with errors" =
+      assert "Only retain input/output data when a run fails" =
                view
                |> element(~s{label#[for="retain_with_errors"]})
                |> render()
@@ -1824,7 +1822,7 @@ defmodule LightningWeb.ProjectLiveTest do
                |> Floki.text()
                |> String.trim()
 
-      assert "Erase all input & output data on workflow completion (zero-persistence)" =
+      assert "Never retain input/output data (zero-persistence)" =
                view
                |> element(~s{label#[for="erase_all"]})
                |> render()
@@ -1848,7 +1846,7 @@ defmodule LightningWeb.ProjectLiveTest do
                |> Floki.parse_fragment!()
                |> Floki.attribute("input", "checked")
 
-      assert "When enabled, you will no longer be able to retry workflow attempts as no data will be stored." =
+      assert "When enabled, you will no longer be able to retry workflow runs as no data will be stored." =
                view
                |> element("#heads-up-description")
                |> render()
@@ -1873,7 +1871,7 @@ defmodule LightningWeb.ProjectLiveTest do
                |> Floki.parse_fragment!()
                |> Floki.attribute("input", "checked")
 
-      assert "When enabled, you will no longer be able to retry workflow attempts as no data will be stored." =
+      assert "When enabled, you will no longer be able to retry workflow runs as no data will be stored." =
                view
                |> element("#heads-up-description")
                |> render()
@@ -1891,13 +1889,13 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, _view, html} =
         live(
           conn,
-          ~p"/projects/#{project.id}/settings#data-retention"
+          ~p"/projects/#{project.id}/settings#data-storage"
         )
 
-      refute html =~ "Workflow Input &amp; Output Data Retention"
+      refute html =~ "Input/Output Data Storage Policy"
 
       assert html =~
-               "Only project owner and admins can view or edit the Data Retention Policy."
+               "Only project owner and admins can view or edit the Data Storage section."
     end
 
     @tag role: :viewer
@@ -1908,13 +1906,13 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, _view, html} =
         live(
           conn,
-          ~p"/projects/#{project.id}/settings#data-retention"
+          ~p"/projects/#{project.id}/settings#data-storage"
         )
 
       refute html =~ "Workflow Input &amp; Output Data Retention"
 
       assert html =~
-               "Only project owner and admins can view or edit the Data Retention Policy."
+               "Only project owner and admins can view or edit the Data Storage section."
     end
 
     @tag role: :admin
@@ -1925,7 +1923,7 @@ defmodule LightningWeb.ProjectLiveTest do
       {:ok, view, _html} =
         live(
           conn,
-          ~p"/projects/#{project.id}/settings#data-retention"
+          ~p"/projects/#{project.id}/settings#data-storage"
         )
 
       # save, navigate to other page and saved option is checked when come back
@@ -1954,7 +1952,7 @@ defmodule LightningWeb.ProjectLiveTest do
             |> render_submit()
 
           assert html =~ "Project updated successfully"
-          assert html =~ "Workflow Input &amp; Output Data Retention"
+          assert html =~ "Input/Output Data Storage Policy"
 
           assert policy ==
                    project.id
@@ -1965,7 +1963,7 @@ defmodule LightningWeb.ProjectLiveTest do
           live(conn, ~p"/projects/#{project.id}/w")
 
           {:ok, view, _html} =
-            live(conn, ~p"/projects/#{project.id}/settings#data-retention")
+            live(conn, ~p"/projects/#{project.id}/settings#data-storage")
 
           assert ["checked"] ==
                    view